Added the comparison function for different models and the comparison function for different versions
